Add mapping of user.ext.eids[] for TDID in Rubicon bidder - refactoring
